cleaned up redux library functions, added foreignAddress toggle to form state and abstacted checkbox into component. Added css to form so that it will remain 30% of screen width
