+++ src/components/w2EmployeeInfo.js
@@ -1,24 +1,24 @@
 import React, { useState } from 'react'
 import { useForm } from 'react-hook-form'
-import { Button, Box, Checkbox, FormControlLabel, Grow } from "@material-ui/core"
+import { Button, Box, Grow } from "@material-ui/core"
 import { Link, useHistory } from "react-router-dom"
 import { useDispatch, useSelector } from 'react-redux'
 
 import locationPostalCodes from './locationPostalCodes'
 import countries from './countries'
 
-import { LabeledInput, LabeledDropdown } from './labeledInput'
+import { LabeledInput, LabeledDropdown, LabeledCheckBox } from './labeledInput'
 import { saveFormData } from '../redux/actions'
 import { getFormData } from '../redux/selectors'
 
 export default function W2EmployeeInfo() {
     const { register, handleSubmit, errors, control } = useForm()
-    const [foreignAddress, setforeignAddress] = useState(false)
     const history = useHistory()
     // const variable dispatch to allow use inside function
     const dispatch = useDispatch()
 
     const prevFormData = useSelector(state => getFormData(state, "W2EmployeeInfo"))
+    const [foreignAddress, setforeignAddress] = useState(prevFormData.foreignAddress === "true")
 
     // component functions
     const onSubmit = formData => {
@@ -26,8 +26,6 @@ export default function W2EmployeeInfo() {
         dispatch(saveFormData(formData, "W2EmployeeInfo"))
         history.push("/")
     }
-    const setForeignAddressFalse = () => setforeignAddress(false)
-    const setForeignAddressTrue = () => setforeignAddress(true)
 
     return (
         <Box display="flex" justifyContent="center">
@@ -62,18 +60,12 @@ export default function W2EmployeeInfo() {
                     <p><strong>Box F - </strong>Employee's Address and Zip Code</p>
                 </Box>
 
-                <Box display="flex" justifyContent="flex-start" paddingTop={1}>
-                    <FormControlLabel
-                        control={<Checkbox checked={!foreignAddress} onChange={setForeignAddressFalse} color="primary" />}
-                        label="No"
-                        ml={0}
-                    />
-                    <FormControlLabel
-                        control={<Checkbox checked={foreignAddress} onChange={setForeignAddressTrue} color="primary" />}
-                        label="Yes"
-                    />
-                    <p>Do you have a foreign address?</p>
-                </Box>
+                <LabeledCheckBox
+                    foreignAddress={foreignAddress}
+                    setforeignAddress={setforeignAddress}
+                    control={control}
+                    description={"Do you have a foreign address?"}
+                />
 
                 <LabeledInput
                     label="Employee's Address" register={register}
