+++ src/components/w2EmployerInfo.js
@@ -1,23 +1,23 @@
 import React, { useState } from 'react'
 import { useForm } from 'react-hook-form'
-import { Button, Box, Checkbox, FormControlLabel, Grow } from "@material-ui/core"
+import { Button, Box, Grow } from "@material-ui/core"
 import { Link, useHistory } from "react-router-dom"
 import { useDispatch, useSelector } from 'react-redux'
 
 import locationPostalCodes from './locationPostalCodes'
 import countries from './countries'
 
-import { LabeledInput, LabeledDropdown} from './labeledInput'
+import { LabeledInput, LabeledDropdown, LabeledCheckBox} from './labeledInput'
 import { saveFormData } from '../redux/actions'
 import { getFormData } from '../redux/selectors'
 
 export default function W2EmployerInfo() {
     const { register, handleSubmit, errors, control } = useForm()
-    const [foreignAddress, setforeignAddress] = useState(false)
     const history = useHistory()
     const dispatch = useDispatch()
 
     const prevFormData = useSelector(state => getFormData(state, "W2EmployerInfo"))
+    const [foreignAddress, setforeignAddress] = useState(prevFormData.foreignAddress === "true")
 
     // component functions
     const onSubmit = formData => { 
@@ -26,8 +26,6 @@ export default function W2EmployerInfo() {
         history.push("w2employeeinfo")
     }
 
-    const setForeignAddressFalse = () => setforeignAddress(false)
-    const setForeignAddressTrue = () => setforeignAddress(true)
     
     return (
         <Box display="flex" justifyContent="center">
@@ -75,18 +73,12 @@ export default function W2EmployerInfo() {
                     errors={errors}
                 />
 
-                <Box display="flex" justifyContent="flex-start" paddingTop={1}>
-                    <FormControlLabel
-                        control={<Checkbox checked={!foreignAddress} onChange={setForeignAddressFalse} color="primary" />}
-                        label="No"
-                        ml={0}
-                    />
-                    <FormControlLabel
-                        control={<Checkbox checked={foreignAddress} onChange={setForeignAddressTrue} color="primary" />}
-                        label="Yes"
-                    />
-                    <p>Do you have a foreign address?</p>
-                </Box>
+                <LabeledCheckBox
+                    foreignAddress={foreignAddress}
+                    setforeignAddress={setforeignAddress}
+                    control={control}
+                    description={"Does your employer have a foreign address?"}
+                />
                 
                 <LabeledInput
                     label="Employer's Address" register={register}
