+++ src/redux/reducer.js
@@ -1,9 +1,5 @@
 import { combineReducers } from 'redux';
 import {
-    ADD_CART_ITEM,
-    REMOVE_CART_ITEM,
-    EMPTY_CART,
-
     SAVE_FORM_DATA
 } from './actions'
 
@@ -19,39 +15,7 @@ function formReducer(state = {}, action) {
     }
 }
 
-function cartReducer(state = {}, action) {
-    switch (action.type) {
-        case ADD_CART_ITEM:
-            return state[action.id] ? {
-                ...state,
-                [action.id]: {
-                    id: action.id,
-                    name: action.name,
-                    price: action.price,
-                    amount: action.amount + state[action.id].amount
-                }
-            } : {
-                    ...state,
-                    [action.id]: {
-                        id: action.id,
-                        name: action.name,
-                        price: action.price,
-                        amount: action.amount
-                    }
-                }
-
-        case REMOVE_CART_ITEM:
-            return (({ [action.id]: _, ...rest }) => rest)(state)
-
-        case EMPTY_CART:
-            return {}
-
-        default:
-            return state;
-    }
-}
 const rootReducer = combineReducers({
     information: formReducer,
-    cart: cartReducer
 })
 export default rootReducer;
