+++ src/redux/selectors.js
@@ -4,11 +4,3 @@ export function getFormData(state, id) {
             {}
         )
 }
-
-export function getProduct(state, id) {
-    return state.products[id]
-}
-
-export function getCartItems(state) {
-    return state.cart
-}
